requirements.txt
mss
pynput
Pillow
PyQt5
fpdf2
